SignNet-BasisNet/requirements.txt
dgl-cu101==0.6.1
gudhi==3.7.1
imageio==2.27.0
matplotlib==3.5.2
networkx==2.6.3
numpy==1.21.6
ogb==1.3.3
pandas==1.3.5
Pillow==9.2.0
pqdm==0.2.0
protobuf==3.20.1
requests==2.28.1
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
tensorboard==2.11.2
tensorboardX==2.5.1
tqdm==4.64.0
urllib3==1.26.10
yacs==0.1.8
